+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/settings/PluginSettings.java
@@ -55,8 +55,8 @@ public class PluginSettings {
     private static final String DEFAULT_CONTENT_CONSUMER = "development_0.0.3_test";
 
     // Default values for IOC Context and Consumer
-    private static final String DEFAULT_IOC_CONTEXT = "ioc_provider_v2";
-    private static final String DEFAULT_IOC_CONSUMER = "iocs_v2";
+    private static final String DEFAULT_IOC_CONTEXT = "ioc_provider_v3";
+    private static final String DEFAULT_IOC_CONSUMER = "iocs_v3";
 
     // Default values for CVE Context and Consumer
     private static final String DEFAULT_CVE_CONTEXT = "vd_1.0.0";
@@ -69,13 +69,13 @@ public class PluginSettings {
     private static PluginSettings INSTANCE;
 
     /** Base Wazuh CTI URL */
-    public static final String CTI_URL = "https://cti.pre.cloud.wazuh.com";
+    public static final String CTI_URL = "https://cti.pre.cloud.wazuh.com/api/v1";
 
     /** The CTI API URL from the configuration file */
     public static final Setting<String> CTI_API_URL =
             Setting.simpleString(
                     "plugins.content_manager.cti.api",
-                    CTI_URL + "/api/v1",
+                    CTI_URL,
                     Setting.Property.NodeScope,
                     Setting.Property.Filtered);