+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/utils/Constants.java
@@ -79,7 +79,6 @@ public class Constants {
     public static final String INDEX_IOCS = ".cti-iocs";
     public static final String INDEX_CVES = ".cti-cves";
     public static final String INDEX_FILTERS = ".engine-filters";
-
     // Resource Types Keys
     public static final String KEY_POLICY = "policy";
     public static final String KEY_INTEGRATIONS = "integrations";